Reduce duplication in range slider components
